requirements.txt
PyYAML>=6.0
Jinja2>=3.1.2
Pillow>=9.0.0
